requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anthropic==0.55.0
    # via mcp-server (pyproject.toml)
anyio==4.9.0
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
arxiv==2.2.0
    # via mcp-server (pyproject.toml)
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
distro==1.9.0
    # via
    #   anthropic
    #   openai
feedparser==6.0.11
    # via arxiv
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
    #   openai
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.10.0
    # via
    #   anthropic
    #   openai
jsonschema==4.24.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
mcp==1.10.1
    # via mcp-server (pyproject.toml)
nest-asyncio==1.6.0
    # via mcp-server (pyproject.toml)
npx==0.1.6
    # via mcp-server (pyproject.toml)
numpy==2.3.1
    # via npx
openai==1.93.0
    # via mcp-server (pyproject.toml)
pydantic==2.11.7
    # via
    #   anthropic
    #   mcp
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
python-dotenv==1.1.1
    # via
    #   mcp-server (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via arxiv
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
sgmllib3k==1.0.0
    # via feedparser
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sse-starlette==2.3.6
    # via mcp
starlette==0.47.1
    # via mcp
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   anthropic
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via mcp
